refactor: deleteFilesByName & improve uploadImage
