+++ src/utils/__test__/deleteAllFilesInDirectory.test.ts
@@ -1,0 +1,43 @@
+import fsMock from 'mock-fs'
+import * as fsWithCallbacks from 'fs'
+
+import { deleteAllFilesInDirectory } from '../deleteAllFilesInDirectory'
+
+const fs = fsWithCallbacks.promises
+
+describe('utils/deleteAllFilesInDirectory', () => {
+  it('delete all the files expect the directories', async () => {
+    fsMock({
+      '/files': {
+        'default.png': '',
+        'user-logo.png': '',
+        'user-logo.jpg': '',
+        directory: {
+          file: ''
+        }
+      }
+    })
+    await deleteAllFilesInDirectory('/files')
+    const directoryContent = await fs.readdir('/files')
+    expect(directoryContent.length).toEqual(1)
+    expect(directoryContent[0]).toEqual('directory')
+  })
+
+  it('delete all the files with all the directories recursively', async () => {
+    fsMock({
+      '/files': {
+        'default.png': '',
+        'user-logo.png': '',
+        'user-logo.jpg': '',
+        directory: {
+          file: ''
+        }
+      }
+    })
+    await deleteAllFilesInDirectory('/files', true)
+    const filesDirectoryContent = await fs.readdir('/files')
+    const directoryContent = await fs.readdir('/files/directory')
+    expect(filesDirectoryContent.length).toEqual(1)
+    expect(directoryContent.length).toEqual(0)
+  })
+})
