+++ src/utils/uploadImage.ts
@@ -1,26 +1,34 @@
 import { UploadedFile } from 'express-fileupload'
+import * as fsWithCallbacks from 'fs'
 import path from 'path'
+import { v4 as uuidv4 } from 'uuid'
 
 import {
   commonErrorsMessages,
   supportedImageMimetype,
   tempPath
 } from './config/constants'
-import { deleteAllFilesInDirectory, deleteFilesByName } from './deleteFiles'
+import { deleteAllFilesInDirectory } from './deleteAllFilesInDirectory'
 import { BadRequestError } from './errors/BadRequestError'
 import { PayloadTooLargeError } from './errors/PayloadTooLargeError'
+
+const fs = fsWithCallbacks.promises
+
+interface UploadImageOptions {
+  image: UploadedFile | UploadedFile[] | undefined
+  propertyName: string
+  oldImage: string
+  imagesPath: string
+}
 
 /**
  * @description Handle upload of an image
  * @returns the complete image name if success otherwise null
  */
-export const uploadImage = async (options: {
-  image: UploadedFile | UploadedFile[] | undefined
-  propertyName: string
-  imageName: string
-  imagesPath: string
-}): Promise<string | null> => {
-  const { image, propertyName, imageName, imagesPath } = options
+export const uploadImage = async (
+  options: UploadImageOptions
+): Promise<string | null> => {
+  const { image, propertyName, oldImage, imagesPath } = options
   if (image != null && !Array.isArray(image)) {
     if (image.truncated) {
       await deleteAllFilesInDirectory(tempPath)
@@ -36,13 +44,13 @@ export const uploadImage = async (options: {
     }
     const splitedMimetype = image.mimetype.split('/')
     const imageExtension = splitedMimetype[1]
-    const completeImageName = `${imageName}.${imageExtension}`
-    // Removes old images
-    await deleteFilesByName({
-      directoryPath: imagesPath,
-      filesNameToDelete: imageName,
-      filesToExclude: ['default.png']
-    })
+    const completeImageName = `${uuidv4()}.${imageExtension}`
+    const oldImagePath = oldImage.split('/')
+    const oldImageName = oldImagePath[oldImagePath.length - 1]
+    if (!oldImageName.startsWith('default')) {
+      // Delete old image
+      await fs.unlink(path.join(imagesPath, oldImageName))
+    }
     await image.mv(path.join(imagesPath, completeImageName))
     await deleteAllFilesInDirectory(tempPath)
     return completeImageName
